Add first pass at project code, tests
* Add project stub, including first unit/integration tests
* Add Docker image build and configuration
* Add pgsql db test container and configuration

The Docker image build has some stuff in it to build a native binary (in the container). Since this project uses the latest version of Vert.x and that has a new version of Netty, I don't have all the quirks worked out (there are some new things that are breaking my previously working configuration). So, the build has a configurable option to build the native binary or the uber-jar. Currently, it's building an uber-jar.