Move DatabaseAccessFailureHandler to AccessLevelHandler::handleFailure
At this point I don't think the code is generic enough for it to be a
standalone class (and thus implying that other classes might want to use
it). For instance, AccessCookieHandler also handles database access
errors, but unlike AccessLevelHandler, it returns HTML responses, so it
can't use DatabaseAccessFailureHandler for them.