Simplify and DRY up error handling code
As I've noted in the comment in MainVerticle, I think it makes sense for
each handler to handle ServiceExceptions on their own. For example, it
makes more sense for AccessModeHandler to treat a NOT_FOUND error in
response to the DB query as HTTP 404, whereas AccessCookieHandler should
treat it as HTTP 400 (or 500 possibly -- see message for 683ba6f). Note,
however, that the code in each handler is a bit simplified (and no
longer in static methods due to the context dependence of each).

The handling of Handlebars template rendering errors was generic enough
for its own class.